Add migration script to help replace deprecated import statements
